Refactor RPC server HTTP request codebase to move it into rpc package instead of osmosis package
